spellcheck fixes

+++ graphrag/api/query.py
@@ -286,7 +286,7 @@ async def multi_global_search(
         nodes_df["title"] = nodes_df["title"].apply(
             lambda x, index_name=index_name: x + f"-{index_name}"
         )
-        max_vals["nodes"] = nodes_df["human_readable_id"].max()
+        max_vals["nodes"] = nodes_df["human_readable_id"].astype(int).max()
         nodes_dfs.append(nodes_df)
 
         # Prepare each index's community reports dataframe for merging
@@ -301,7 +301,7 @@ async def multi_global_search(
             }
         community_reports_df["community"] += max_vals["community_reports"] + 1
         community_reports_df["human_readable_id"] += max_vals["community_reports"] + 1
-        max_vals["community_reports"] = community_reports_df["community"].max()
+        max_vals["community_reports"] = community_reports_df["community"].astype(int).max()
         community_reports_dfs.append(community_reports_df)
 
         # Prepare each index's communities dataframe for merging
@@ -318,7 +318,7 @@ async def multi_global_search(
             lambda x: x if x == -1 else x + max_vals["community"] + 1
         )
         communities_df["human_readable_id"] += max_vals["community"] + 1
-        max_vals["community"] = communities_df["community"].max()
+        max_vals["community"] = communities_df["community"].astype(int).max()
         communities_dfs.append(communities_df)
 
         # Prepare each index's entities dataframe for merging
@@ -335,7 +335,7 @@ async def multi_global_search(
         entities_df["text_unit_ids"] = entities_df["text_unit_ids"].apply(
             lambda x, index_name=index_name: [i + f"-{index_name}" for i in x]
         )
-        max_vals["entities"] = entities_df["human_readable_id"].max()
+        max_vals["entities"] = entities_df["human_readable_id"].astype(int).max()
         entities_dfs.append(entities_df)
 
     # Merge the dataframes
@@ -614,7 +614,7 @@ async def multi_local_search(
         nodes_df["id"] = nodes_df["id"].apply(
             lambda x, index_name=index_name: x + f"-{index_name}"
         )
-        max_vals["nodes"] = nodes_df["human_readable_id"].max()
+        max_vals["nodes"] = nodes_df["human_readable_id"].astype(int).max()
         nodes_dfs.append(nodes_df)
 
         # Prepare each index's community reports dataframe for merging
@@ -629,7 +629,7 @@ async def multi_local_search(
             }
         community_reports_df["community"] += max_vals["community_reports"] + 1
         community_reports_df["human_readable_id"] += max_vals["community_reports"] + 1
-        max_vals["community_reports"] = community_reports_df["community"].max()
+        max_vals["community_reports"] = community_reports_df["community"].astype(int).max()
         community_reports_dfs.append(community_reports_df)
 
         # Prepare each index's entities dataframe for merging
@@ -649,7 +649,7 @@ async def multi_local_search(
         entities_df["text_unit_ids"] = entities_df["text_unit_ids"].apply(
             lambda x, index_name=index_name: [i + f"-{index_name}" for i in x]
         )
-        max_vals["entities"] = entities_df["human_readable_id"].max()
+        max_vals["entities"] = entities_df["human_readable_id"].astype(int).max()
         entities_dfs.append(entities_df)
 
         # Prepare each index's relationships dataframe for merging
